% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{bdg.estimate}
\alias{bdg.estimate}
\title{Wrapper for estimating posterior probabilities with bdgraph}
\usage{
bdg.estimate(data, mpl = FALSE, ...)
}
\arguments{
\item{data}{An n x p matrix of data}

\item{mpl}{Use true for marginal pseudo-likehlihood}

\item{...}{One or more arguments passed to bdgraph}
}
\value{
A bdgraph object
}
\description{
Wrapper for estimating posterior probabilities with bdgraph
}
