% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{bdg.metanet}
\alias{bdg.metanet}
\title{Estimate a meta-network from module eigengenes}
\usage{
bdg.metanet(eigs, cut = 0.5, mpl = FALSE, ...)
}
\arguments{
\item{eigs}{Module eigengenes}

\item{cut}{Threshold for selected edges from posterior probabilities}

\item{mpl}{Use true for marginal pseudo-likehlihood}

\item{...}{One or more arguments passed to bdgraph}
}
\value{
A list of data pertaining to resulting meta-network
}
\description{
Estimate a meta-network from module eigengenes
}
