% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiClass.R
\name{BiClass_Shine}
\alias{BiClass_Shine}
\title{BiClass_Shine}
\usage{
BiClass_Shine(times = 1001, logistic = TRUE, RandomForest = TRUE, SVM = TRUE)
}
\arguments{
\item{times}{the loop times in sample selection,default is 1001,and the number must be odd number.}

\item{logistic}{do logistic regression}

\item{RandomForest}{do RandomForest}

\item{SVM}{do support vector machine}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can do classify analysis using logistic,svm and randomforest
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
