% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{blanket.add.modpairs}
\alias{blanket.add.modpairs}
\title{Lifts the blanket from pairs of modules}
\usage{
blanket.add.modpairs(blanket, mods, modpairs)
}
\arguments{
\item{blanket}{A blanket matrix}

\item{mods}{A list of co-expression modules}

\item{modpairs}{A dataframe of pairs of co-expression modules to fill in}
}
\value{
A blanket matrix
}
\description{
Lifts the blanket from pairs of modules
}
