% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{blanket.add.mods}
\alias{blanket.add.mods}
\title{Lifts the blanket from single modules}
\usage{
blanket.add.mods(blanket, mods, val = 0.5)
}
\arguments{
\item{blanket}{A blanket matrix}

\item{mods}{A list of co-expression modules to fill in}
}
\value{
A blanket matrix
}
\description{
Lifts the blanket from single modules
}
