% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{blanket.cred}
\alias{blanket.cred}
\title{Computes complexity reduction of the blanket}
\usage{
blanket.cred(blanket)
}
\arguments{
\item{blanket}{A blanket matrix}
}
\value{
Fold complexity reduction
}
\description{
Computes complexity reduction of the blanket
}
