% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{blanket.inform}
\alias{blanket.inform}
\title{Adds prior information where the blanket is lifted}
\usage{
blanket.inform(blanket, prior)
}
\arguments{
\item{blanket}{A blanket matrix}

\item{prior}{A matrix of probabilities with the same dimensions as the blanket}
}
\value{
A blanket matrix
}
\description{
Adds prior information where the blanket is lifted
}
