% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{blanket.lift}
\alias{blanket.lift}
\title{Lifts the blanket from within or between modules}
\usage{
blanket.lift(blanket, mods, pairs = NULL, val = 0.5)
}
\arguments{
\item{blanket}{A blanket matrix}

\item{mods}{A list of co-expression modules to fill in}

\item{pairs}{A dataframe of pairs of co-expression modules to fill in}

\item{val}{Value for lifted links}
}
\value{
A blanket matrix
}
\description{
Lifts the blanket from within or between modules
}
