% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{blanket.new}
\alias{blanket.new}
\title{Create a new blanket covering the entire graph search space}
\usage{
blanket.new(genes, val = 0)
}
\arguments{
\item{genes}{The genes of the graph}
}
\value{
A gene by gene matrix filled with zeroes
}
\description{
Create a new blanket covering the entire graph search space
}
