% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{BoxPlot}
\alias{BoxPlot}
\title{BoxPlot}
\usage{
BoxPlot(metabolite = NULL)
}
\arguments{
\item{metabolite}{the metabolite to build boxplot}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
