% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow.R
\name{build.workflow}
\alias{build.workflow}
\title{Generate a nextflow workflow}
\usage{
build.workflow(
  hierarchy,
  condition,
  path.eset,
  path.genes,
  path.blanket = NULL,
  outdir = ".",
  iters = 5000,
  cores = 1
)
}
\arguments{
\item{hierarchy}{A string describing the hierarchy in dot language notation}

\item{condition}{The column in pData where subtypes are described}

\item{path.eset}{A full path to an expression set object saved as an rds file}

\item{path.genes}{A full path to a character vector of genes saved as an rds file}

\item{path.blanket}{A full path to a gene x gene constraint matrix saved as an rds file}

\item{outdir}{A directory where the workflow will be generated}

\item{iters}{The number of iterations for the sampling algorithm}

\item{cores}{The number of cores for each network to use for parallel execution}
}
\description{
Generate a nextflow workflow
}
