% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{Calculate_ppm}
\alias{Calculate_ppm}
\title{Calculate_ppm}
\usage{
Calculate_ppm()
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to Calculate_ppm
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
