% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{check.sft}
\alias{check.sft}
\title{Check and choose a soft threshold}
\usage{
check.sft(sft)
}
\arguments{
\item{sft}{Table of values for scale-free fit}
}
\value{
A theshold
}
\description{
Check and choose a soft threshold
}
\keyword{internal}
