% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicular.R
\name{CircularPlot}
\alias{CircularPlot}
\title{CircularPlot}
\usage{
CircularPlot(label_size = 5)
}
\arguments{
\item{label_size}{5}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw CircularPlot
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
