% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Column_Rows.R
\name{Column_Rows}
\alias{Column_Rows}
\title{Column_Rows}
\usage{
Column_Rows(dir = "/Users/data.csv")
}
\arguments{
\item{dir}{data path}
}
\description{
print cols and rows
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
