% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorAnalysis.R
\name{CorAnalysis}
\alias{CorAnalysis}
\title{CorAnalysis}
\usage{
CorAnalysis(data)
}
\arguments{
\item{data}{a dataframe to do correlation analysis.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can do correlation analysis
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
