% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_theme.R
\name{dashboard_plotly_theme}
\alias{dashboard_plotly_theme}
\title{Dashboard plotly theme}
\usage{
dashboard_plotly_theme(pp, ...)
}
\arguments{
\item{pp}{A plotly object.}

\item{...}{Additional arguments passed to `plotly::layout()`.}
}
\value{
A plotly object.
}
\description{
Applies consistent hover styling and removes the plotly modebar except for
the download-to-image button.
}
