% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{demo data for shine}
\format{A data frame with 1351 rows and 181 variables:
\describe{
  \item{name}{compound name or feature name}
  \item{mz}{mz}
  \item{rt}{rt}
  \item{isotope}{isotope}
  \item{QC1}{QC}
  \item{QC2}{QC}
  \item{QC3}{QC}
  \item{QC4}{QC}
  \item{QC5}{QC}
  \item{S20}{S}
  \item{S30}{S}
  \item{S39}{S}
  \item{S49}{S}
  \item{S59}{S}
  \item{S69}{S}
  \item{S78}{S}
  \item{S2}{S}
  \item{S3}{S}
  \item{S4}{S}
}}
\usage{
data
}
\description{
A dataset containing the name mz rt and sample intensity:
}
\keyword{datasets}
