% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{data_filter}
\alias{data_filter}
\title{data_filter}
\usage{
data_filter(p = 0.05, comp1 = 1, comp2 = 1, fc_toplimit = 1.2,
  fc_lowerlimit = 0.8)
}
\arguments{
\item{p}{p value}

\item{comp1}{component 1 in PLSDA}

\item{comp2}{component 2 in PLSDA}

\item{fc_toplimit}{top limit of fold change}

\item{fc_lowerlimit}{lower limit of fold change}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to filter p value more than 0.05 or any value you can choose,
foldchange less than 1.2 and more than 0.8 or any value you can choose,
and vip less than 1 or any value you can choose
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
