% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFilter.R
\name{DataFilter_Shine}
\alias{DataFilter_Shine}
\title{DataFilter_Shine}
\usage{
DataFilter_Shine(
  p = 0.05,
  comp1 = 1,
  comp2 = 1,
  fc_toplimit = 4/3,
  fc_lowerlimit = 3/4
)
}
\arguments{
\item{p}{p value}

\item{comp1}{component 1 in PLSDA}

\item{comp2}{component 2 in PLSDA}

\item{fc_toplimit}{top limit of fold change}

\item{fc_lowerlimit}{lower limit of fold change}
}
\value{
filter p,foldchange,vip value that do not match the condition .
}
\description{
a function to filter p value more than 0.05 or any value you can choose,
foldchange less than 1.2 and more than 0.8 or any value you can choose,
and vip less than 1 or any value you can choose
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
