% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{eig.exp}
\alias{eig.exp}
\title{Variance explained by eigengen from the kth principal component for a module}
\usage{
eig.exp(dat, mod, k = 1)
}
\arguments{
\item{dat}{Expression data}

\item{mod}{A module of genes}

\item{k}{The principal component to use}
}
\value{
Variance explained by the kth principle component
}
\description{
Variance explained by eigengen from the kth principal component for a module
}
\keyword{internal}
