% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{eig.get}
\alias{eig.get}
\title{Module eigengene from the kth principal component for a module}
\usage{
eig.get(dat, mod, k = 1)
}
\arguments{
\item{dat}{Expression data}

\item{mod}{A module of genes}

\item{k}{The principal component to use}
}
\value{
An eigengene for the kth principle component
}
\description{
Module eigengene from the kth principal component for a module
}
\keyword{internal}
