% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureAnalysis.R
\name{FeatureAnalysis}
\alias{FeatureAnalysis}
\title{FeatureAnalysis}
\usage{
FeatureAnalysis(data = NULL, sample.info = NULL)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns..}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate mz vs rt plot and QC distribution plot,
also can filter isotope,rsd and zero value.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
data(data, package = "Shine")
data(sample.info, package = "Shine")
##create a folder for demo
dir.create("demo")
setwd("demo")
# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample_info, "sample_info.csv", row.names = FALSE)
# Analysis process
FeatureAnalysis(data = "data.csv",sample.info = "sample.info.csv")
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
