% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter.var}
\alias{filter.var}
\title{Filter variables with zero variance in one or more subtypes}
\usage{
\method{filter}{var}(eset, column, subtypes, fn = var)
}
\arguments{
\item{eset}{An expression set object}

\item{column}{The column in pData where subtypes are described}

\item{subtypes}{One or more unique subtypes}

\item{fn}{A function required to be non-zero}
}
\value{
A vector of filtered genes
}
\description{
Filter variables with zero variance in one or more subtypes
}
