% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestAnalysis.R
\name{ForestAnalysis_Shine}
\alias{ForestAnalysis_Shine}
\title{ForestAnalysis_Shine}
\usage{
ForestAnalysis_Shine(group = c("Gout", "Normal"))
}
\arguments{
\item{group}{group set.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can do forest analysis
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
