% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{fuzzy.mods}
\alias{fuzzy.mods}
\title{Classify fuzzy modules with quadratic disciminant analysis}
\usage{
fuzzy.mods(dat, mods, cores = 1, p = 0.5, fn = bicor)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{cores}{The number of cores for parallel execution}

\item{p}{Posterior probability threshold}

\item{fn}{A correlation function}
}
\value{
A list of fuzzy modules
}
\description{
Classify fuzzy modules with quadratic disciminant analysis
}
