% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{fuzzy.plot}
\alias{fuzzy.plot}
\title{Visualize fuzzy module classification}
\usage{
fuzzy.plot(dat, mods, mod, p = 0.5, fn = bicor, size = 2.5, resolution = 100)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{mod}{A module name}

\item{p}{Posterior probability threshold}

\item{fn}{A correlation function}

\item{size}{Size of data points}

\item{resolution}{Resolution of heatmap}
}
\value{
A ggplot object
}
\description{
Visualize fuzzy module classification
}
