% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{fuzzy.plots}
\alias{fuzzy.plots}
\title{Visualize multiple fuzzy module classifications}
\usage{
fuzzy.plots(dat, mods, ...)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{...}{Additional arguments passed to /code{shine::fuzzy.plot}}
}
\value{
A ggplot object
}
\description{
Visualize multiple fuzzy module classifications
}
