% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{fuzzy.predict}
\alias{fuzzy.predict}
\title{Classify fuzzy modules with quadratic discriminant analysis}
\usage{
fuzzy.predict(dat, mods, mod, fn = bicor)
}
\description{
Classify fuzzy modules with quadratic discriminant analysis
}
