% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not-in operator}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{Vector of values to test.}

\item{table}{Vector of values to test against.}
}
\value{
Logical vector.
}
\description{
Not-in operator
}
