% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{graph.similarity}
\alias{graph.similarity}
\title{Szymkiewicz–Simpson coefficient of edges}
\usage{
graph.similarity(ig1, ig2)
}
\arguments{
\item{ig1}{An igraph object}

\item{ig2}{An igraph object}
}
\value{
An edge similarity
}
\description{
Szymkiewicz–Simpson coefficient of edges
}
