% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedbox.R
\name{Grouped_Box}
\alias{Grouped_Box}
\title{Grouped_Box}
\usage{
Grouped_Box(
  font_size = 20,
  levels = c("M1", "M2", "M3"),
  zscore = TRUE,
  ylab = "Relative Abundance (log2)",
  info_name = "info.csv",
  data_name = "data.csv"
)
}
\arguments{
\item{font_size}{font_size}

\item{levels}{order of group}

\item{zscore}{default TRUE}

\item{ylab}{name}

\item{info_name}{info_name}

\item{data_name}{data_name}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
