% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_box.R
\name{Grouped_BoxVilion}
\alias{Grouped_BoxVilion}
\title{Grouped_BoxVilion}
\usage{
Grouped_BoxVilion(
  stat_type = "nonparametric",
  data_name = "data.csv",
  info_name = "info.csv",
  plot.type = "box",
  zscore = TRUE,
  palette = "nrc_npg",
  plot_nrow = 2,
  pairwise.display = "significant",
  p.adjust.method = "fdr",
  ylab = "Relative Abundance (log2)",
  levels = c("M1", "M2", "M3")
)
}
\arguments{
\item{stat_type}{parametric}

\item{data_name}{data.csv}

\item{info_name}{info_name}

\item{plot.type}{box}

\item{zscore}{default TRUE}

\item{palette}{palette}

\item{plot_nrow}{plot_nrow}

\item{pairwise.display}{significant}

\item{p.adjust.method}{fdr}

\item{ylab}{name}

\item{levels}{order of group}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
