% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h}
\alias{h}
\title{pathway data for human}
\format{A list contain pathway information:
\describe{
  \item{hsa}{ID}
}}
\usage{
h
}
\description{
A dataset containing the pathway information:
}
\keyword{datasets}
