% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Shine.R
\name{Heatmap_Complex}
\alias{Heatmap_Complex}
\title{Heatmap}
\usage{
Heatmap_Complex(
  data = "data.csv",
  colour = c("navy", "white", "firebrick3"),
  a = -2,
  b = 2,
  width = 12,
  height = 7,
  QC = TRUE,
  scale_row = TRUE,
  fontsize = 10,
  show_colnames = FALSE,
  show_rownames = FALSE,
  cluster_cols = FALSE,
  group = c("case", "control")
)
}
\arguments{
\item{data}{data name}

\item{colour}{colour selection.}

\item{a}{lowerlimt of colour.}

\item{b}{toplimit of colour}

\item{width}{12}

\item{height}{7}

\item{QC}{QC exist}

\item{scale_row}{scale or not}

\item{fontsize}{the font size}

\item{show_colnames}{show_colnames}

\item{show_rownames}{show_rownames}

\item{cluster_cols}{default is FALSE}

\item{group}{group info.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate heatmap
}
\author{
Xia Shen
\email{shenxia@shanghaitech.edu.cn}
}
