% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.kegg.pathway.R
\docType{data}
\name{hsa.kegg.pathway}
\alias{hsa.kegg.pathway}
\title{pathway data for human}
\format{
A list contain pathway information:
\describe{
  \item{hsa.kegg.pathway}{ID of all compounds in hsa.kegg.pathway}
}
}
\usage{
hsa.kegg.pathway
}
\description{
A dataset containing the pathway information:
}
\keyword{datasets}
