% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.compound.ID.R
\docType{data}
\name{hsa_compound_ID}
\alias{hsa_compound_ID}
\title{the pathway compound data for human compound}
\format{A dataframe with 3027 rows and 3 variables:
\describe{
  \item{hsa_compound_data1}{ID of all compounds in hsa.kegg.pathway}
  ...

}}
\usage{
hsa_compound_ID
}
\description{
A dataset containing the pathway ID information:
}
\keyword{datasets}
