% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMatch.R
\name{iMatch}
\alias{iMatch}
\title{iMatch}
\usage{
iMatch(name = "name")
}
\arguments{
\item{name}{the column which you want to match}
}
\value{
get data you want.
}
\description{
a function to mtach data.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
