% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISboxplot.R
\name{ISboxplot}
\alias{ISboxplot}
\title{ISboxplot}
\usage{
ISboxplot(sample = NULL, width = 10, height = 6)
}
\arguments{
\item{sample}{a numeric object show the number of sample.}

\item{width}{the width of plot}

\item{height}{the height of plot}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate internal standard boxplot
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
