% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsotopeFilter.R
\name{IsotopeFilter}
\alias{IsotopeFilter}
\title{IsotopeFilter}
\usage{
IsotopeFilter()
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can filter isotope.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
