% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{keep.var}
\alias{keep.var}
\title{Keep variables with non-zero variance within subtypes}
\usage{
keep.var(eset, column, subtypes, fn = var)
}
\arguments{
\item{eset}{An expression set object}

\item{column}{The column in pData where subtypes are described}

\item{subtypes}{One or more unique subtypes}

\item{fn}{A function required to be non-zero}
}
\value{
A vector of genes
}
\description{
Keep variables with non-zero variance within subtypes
}
