% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph_cumulative.R
\name{line_graph_cumulative}
\alias{line_graph_cumulative}
\title{Cumulative line chart}
\usage{
line_graph_cumulative(data, display, xaxis, color, facet, unit = "units")
}
\arguments{
\item{data}{A data frame.}

\item{display}{One of "numbers" or "percent".}

\item{xaxis}{Column name for the date x-axis.}

\item{color}{Column name for line color grouping, or "none".}

\item{facet}{Column name for faceting, or "none".}

\item{unit}{Label for count units shown in tooltips.}
}
\value{
A plotly object.
}
\description{
Cumulative line chart
}
