% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_s3.R
\name{list_bucket}
\alias{list_bucket}
\title{List Contents of an S3 Bucket}
\usage{
list_bucket(bucket)
}
\arguments{
\item{bucket}{Character string specifying the S3 bucket name}
}
\value{
A tibble containing information about files in the bucket
}
\description{
List Contents of an S3 Bucket
}
