% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m}
\alias{m}
\title{pathway data for mouse}
\format{A list contain pathway information:
\describe{
  \item{mmu}{ID}
}}
\usage{
m
}
\description{
A dataset containing the pathway information:
}
\keyword{datasets}
