% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{me.get}
\alias{me.get}
\title{Module eigengenes from the kth principal component for all modules}
\usage{
me.get(dat, mods, k = 1)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{k}{The principal component to use}
}
\value{
A matrix of module eigengenes
}
\description{
Module eigengenes from the kth principal component for all modules
}
