% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabolites_Classify.R
\name{Metabolites_Classify}
\alias{Metabolites_Classify}
\title{Metabolites_Classify}
\usage{
Metabolites_Classify(
  data_name = "data for classify.csv",
  specias_pathway_database = c(hsa.kegg.pathway, mmu.kegg.pathway),
  specias_compound_database = c(hsa_compound_ID, mmu_compound_ID)
)
}
\arguments{
\item{data_name}{the data_name}

\item{specias_pathway_database}{the specia of sample}

\item{specias_compound_database}{the specia of sample}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function Metabolites_Classify.
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
