% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{metanet.build}
\alias{metanet.build}
\title{Build a meta-network from module eigengenes}
\usage{
metanet.build(eigs, cut = 0.5, mpl = FALSE, ...)
}
\arguments{
\item{eigs}{Module eigengenes}

\item{cut}{Threshold for selected edges from posterior probabilities}

\item{mpl}{Use true for marginal pseudo-likehlihood}

\item{...}{One or morge arguments passed to bdgraph}
}
\value{
A list of data pertaining to resulting meta-network
}
\description{
Build a meta-network from module eigengenes
}
