% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{mm.get}
\alias{mm.get}
\title{Module memberships by correlation with eigengene for a module}
\usage{
mm.get(dat, mods, k = 1, fn = bicor)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{k}{The principal component to use}

\item{fn}{A correlation function}
}
\value{
A matrix of module memberships
}
\description{
Module memberships by correlation with eigengene for a module
}
