% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{mm.merged}
\alias{mm.merged}
\title{Wrapper to merge module memberships for the first two eigengenes}
\usage{
mm.merged(dat, mods, fn = bicor)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{fn}{A correlation function}
}
\value{
A dataframe of module memberships
}
\description{
Wrapper to merge module memberships for the first two eigengenes
}
