% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmu.kegg.pathway.R
\docType{data}
\name{mmu.kegg.pathway}
\alias{mmu.kegg.pathway}
\title{pathway data for mouse}
\format{A list contain pathway information:
\describe{
  \item{mmu.kegg.pathway}{ID of all compounds in mmu.kegg.pathway}
}}
\usage{
mmu.kegg.pathway
}
\description{
A dataset containing the pathway information:
}
\keyword{datasets}
