% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{mod.plot}
\alias{mod.plot}
\title{Plot membership of all genes for one module}
\usage{
mod.plot(dat, mods, mod, colors, size = 1, fn = bicor)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{mod}{A module name}

\item{colors}{Gene colors}

\item{size}{Size of data points}

\item{fn}{A correlation function}
}
\value{
A ggplot object
}
\description{
Plot membership of all genes for one module
}
