% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.layout}
\alias{model.layout}
\title{Spring-embedded layout}
\usage{
model.layout(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
A graph layout_with_graphopt
}
\description{
Spring-embedded layout
}
\keyword{internal}
