% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.lfr}
\alias{model.lfr}
\title{Simulate network via lancichinetti–fortunato–radicchi benchmark}
\usage{
model.lfr(
  n = 100,
  tau1 = 3,
  tau2 = 2,
  mu = 0.1,
  average_degree = 5,
  min_community = 5,
  max_community = 30,
  seed = 1,
  rpy = NULL,
  ...
)
}
\arguments{
\item{n}{Number of nodes in the created graph}

\item{tau1}{Power law exponent for the degree distribution [> 1]}

\item{tau2}{Power law exponent for the community size distribution in the created graph [> 1]}

\item{mu}{Fraction of intra-community edges incident to each node [0, 1]}

\item{average_degree}{Average degree of nodes [0, n]}

\item{min_community}{Minimum size of communities in the graph}

\item{max_community}{Maximum size of communities in the graph}

\item{seed}{Set a seed}

\item{rpy}{Optionally set a specific python path for reticulate to use}

\item{...}{Additional arguments passed to /code{LFR}}
}
\value{
An igraph object
}
\description{
Simulate network via lancichinetti–fortunato–radicchi benchmark
}
