% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.mpa}
\alias{model.mpa}
\title{Simulate network via modular preferential attachment model}
\usage{
model.mpa(
  p = 300,
  m = 6,
  q.hub = 0.95,
  m.links = 6,
  power = 1.7,
  z.appeal = 1,
  ...
)
}
\description{
Simulate network via modular preferential attachment model
}
