% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.pa}
\alias{model.pa}
\title{Simulate network via preferential attachment model}
\usage{
model.pa(p = 300, power = 1, z.appeal = 1, ...)
}
\description{
Simulate network via preferential attachment model
}
