% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.rename}
\alias{model.rename}
\title{Rename igraph vertices as characters}
\usage{
model.rename(ig, prefix = "G")
}
\arguments{
\item{ig}{An igraph object}

\item{prefix}{A prefix for vertex labels}
}
\value{
An igraph object
}
\description{
Rename igraph vertices as characters
}
\keyword{internal}
