% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.sim}
\alias{model.sim}
\title{Simulate multivariate Gaussian data for a model}
\usage{
model.sim(n, ig, seed = 1)
}
\arguments{
\item{n}{The number of samples to generate data for}

\item{ig}{An igraph object for the graph structure}

\item{seed}{A number to seed the random data}
}
\value{
The true graph structure and multivariate gaussian data
}
\description{
Simulate multivariate Gaussian data for a model
}
