% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.simplify}
\alias{model.simplify}
\title{Removes loops, multi-edges, and isolated vertices from graph}
\usage{
model.simplify(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
An igraph object
}
\description{
Removes loops, multi-edges, and isolated vertices from graph
}
\keyword{internal}
