% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models.plot}
\alias{models.plot}
\title{Plot one or more models highlighting differences}
\usage{
models.plot(igs)
}
\arguments{
\item{igs}{A list of igraph objects}
}
\value{
A graph visualization
}
\description{
Plot one or more models highlighting differences
}
