% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models.similarity}
\alias{models.similarity}
\title{Edge similarity of two or more graphs}
\usage{
models.similarity(igs)
}
\arguments{
\item{igs}{A list of igraph objects}
}
\value{
An edge similarity
}
\description{
Edge similarity of two or more graphs
}
