% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUVR_ANALYSIS.R
\name{MUVR_ANALYSIS}
\alias{MUVR_ANALYSIS}
\title{MUVR_ANALYSIS}
\usage{
MUVR_ANALYSIS(group = "Normal", data_position = 1, info_position = 2)
}
\arguments{
\item{group}{group}

\item{data_position}{1}

\item{info_position}{2}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to do MUVR_ANALYSIS
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
