% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAfilter.R
\name{NAfilter}
\alias{NAfilter}
\title{NA filter}
\usage{
NAfilter(data)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}
}
\description{
NA filter
}
