% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneShot.R
\name{OneShot}
\alias{OneShot}
\title{OneShot}
\usage{
OneShot(path1 = NULL, path2 = NULL, path3 = NULL, group = c("case",
  "control"), pathway.enrichment = TRUE, correct = TRUE)
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to do all work
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
