% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLS_Ana.R
\name{OPLSDA}
\alias{OPLSDA}
\title{OPLSDA}
\usage{
OPLSDA(group = "Gout", data_position = 1, scale = "ParetoNorm",
  repeats = 200)
}
\arguments{
\item{group}{group}

\item{data_position}{1}

\item{scale}{method to scale data}

\item{repeats}{200}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to do OPLSDA
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
