% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{PathwayAnalysis}
\alias{PathwayAnalysis}
\title{PathwayAnalysis}
\usage{
PathwayAnalysis(font_size = 15, path_font_size = 5, nrow = 5)
}
\arguments{
\item{font_size}{15}

\item{path_font_size}{5}

\item{nrow}{5}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to do PathwayAnalysis
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
