% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayBarplot.R
\name{PathwayBarplot}
\alias{PathwayBarplot}
\title{pathway enrichment barplot}
\usage{
PathwayBarplot(font_size = 26)
}
\arguments{
\item{font_size}{the font_size of plot,default is 26.}
}
\description{
pathway enrichment barplot
}
