% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayEnrich.R
\name{PathwayEnrich}
\alias{PathwayEnrich}
\title{PathwayEnrich}
\usage{
PathwayEnrich(
  specias_pathway_database = c(hsa.kegg.pathway, mmu.kegg.pathway),
  specias_compound_database = c(hsa_compound_ID, mmu_compound_ID),
  font_size = 20
)
}
\arguments{
\item{specias_pathway_database}{the specia of sample}

\item{specias_compound_database}{the specia of sample}

\item{font_size}{20}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to do PathwayAnalysis
}
\author{
Xia Shen
\email{qq951633542@163.com}
}
