% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA}
\usage{
PCA(data = NULL, sample.info = NULL, ind = FALSE, ellipse = FALSE,
  both = FALSE, neither = TRUE)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{ind}{default is FALSE.}

\item{ellipse}{default is FALSE.}

\item{both}{default is FALSE.}

\item{neither}{default is TRUE.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
