% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation_Shine.R
\name{Permutation_Shine}
\alias{Permutation_Shine}
\title{Permutation_Shine}
\usage{
Permutation_Shine(repeats = 200, ncomp = 2, group = c("case",
  "control"), scale = c("UV", "pareto"))
}
\arguments{
\item{repeats}{the repeats to do,default is 200}

\item{ncomp}{the components of pls}

\item{group}{the group of the data}

\item{scale}{scale method}
}
\value{
get data you want.
}
\description{
a function to do permutation test.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
