% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{plt.graph}
\alias{plt.graph}
\title{Plot an igraph object}
\usage{
plt.graph(ig, seed = 1)
}
\arguments{
\item{ig}{An igraph object}

\item{seed}{A number to seed random layouts}
}
\value{
A graph visualization
}
\description{
Plot an igraph object
}
