% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{plt.graphs}
\alias{plt.graphs}
\title{Plot multiple igraph objects}
\usage{
plt.graphs(...)
}
\arguments{
\item{...}{One or more igraph objects}
}
\value{
A graph visualization
}
\description{
Plot multiple igraph objects
}
