% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow.R
\name{plt.hierarchy}
\alias{plt.hierarchy}
\title{Plot a hierarchy description}
\usage{
plt.hierarchy(hierarchy)
}
\arguments{
\item{hierarchy}{A string describing the hierarchy in dot language notation}
}
\description{
Plot a hierarchy description
}
