% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{plt.soft}
\alias{plt.soft}
\title{Plot curve for soft thresholding}
\usage{
plt.soft(sft, powers)
}
\arguments{
\item{sft}{Table of values for scale-free fit}

\item{powers}{The power vector tested}
}
\value{
A plot
}
\description{
Plot curve for soft thresholding
}
\keyword{internal}
