% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{prior.soften}
\alias{prior.soften}
\title{Removes absolute values in posterior before being used as a prior}
\usage{
prior.soften(prior)
}
\arguments{
\item{prior}{A matrix of probabilities}
}
\value{
A matrix of probabilities
}
\description{
Removes absolute values in posterior before being used as a prior
}
