% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data.R
\name{qc_data}
\alias{qc_data}
\title{Data Quality Control Function}
\usage{
qc_data(data, required_columns = NULL, alert = shiny::isRunning())
}
\arguments{
\item{data}{A data frame to validate.}

\item{required_columns}{A character vector of column names that must exist in the data. Default is NULL (no required columns).}

\item{alert}{Logical. If TRUE, displays an alert using shinyWidgets if a validation fails. Defaults to the result of `shiny::isRunning()`.}
}
\value{
Logical. Returns TRUE if the data frame passes all checks; otherwise, FALSE.
}
\description{
Validates the provided data frame by checking its format and ensuring that required columns exist.
}
\examples{
\dontrun{
df <- data.frame(a = 1:5, b = 6:10)
qc_data(df, required_columns = c("a", "b"))
qc_data(df, required_columns = c("a", "c"))
}

}
