% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.info}
\alias{sample.info}
\title{demo data for shine}
\format{A data frame with 177 rows and 5 variables:
\describe{
  \item{sample.name}{compound name or feature name}
  \item{injection.order}{order}
  \item{class}{class}
  \item{batch}{batch}
  \item{group}{group}

}}
\usage{
sample.info
}
\description{
A dataset containing the sample.name injection.order batch,group and class:
}
\keyword{datasets}
