% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter3D.R
\name{Scatter3D}
\alias{Scatter3D}
\title{Scatter3D}
\usage{
Scatter3D()
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw 3D scatter plot
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
