% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{sft.check}
\alias{sft.check}
\title{Check and choose a soft threshold}
\usage{
sft.check(sft)
}
\arguments{
\item{sft}{Table of values for scale-free fit}
}
\value{
A theshold
}
\description{
Check and choose a soft threshold
}
\keyword{internal}
