% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{sft.plot}
\alias{sft.plot}
\title{Plot curve for soft thresholding}
\usage{
sft.plot(sft, powers)
}
\arguments{
\item{sft}{Table of values for scale-free fit}

\item{powers}{The power vector tested}
}
\value{
A plot
}
\description{
Plot curve for soft thresholding
}
\keyword{internal}
