% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sim.graph}
\alias{sim.graph}
\title{Barabasi-Albert algorithm for scale-free graphs}
\usage{
sim.graph(p, seed = NULL, ...)
}
\arguments{
\item{p}{The number of vertices}

\item{seed}{The starting igraph for the preferential attachment algorithm}

\item{...}{One or morge arguments passed to the algorithm}
}
\value{
An igraph object
}
\description{
Barabasi-Albert algorithm for scale-free graphs
}
