% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim.graphs}
\alias{sim.graphs}
\title{Simulate a range of scale-free graphs of different similarity}
\usage{
sim.graphs(
  n,
  p.graphs,
  p.start = p.graphs/2,
  p.step = (p.start - p.graphs)/10,
  ...
)
}
\arguments{
\item{n}{The number of graphs}

\item{p.graphs}{The number of desired vertices}

\item{p.start}{The number of vertices in the seed graph for the least similar graphs}

\item{p.step}{The step size between the seed graphs for the least and most similar grpahs}
}
\value{
A list of graphs with decreasing similarity
}
\description{
Simulate a range of scale-free graphs of different similarity
}
