% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaAnalysis.R
\name{StaAnalysis}
\alias{StaAnalysis}
\title{StaAnalysis}
\usage{
StaAnalysis(p.cutoff = 0, group = c("case", "control"),
  splot = FALSE, unitest = c("t.test", "wilcox.test"),
  pcorrect = TRUE, xlim = c(-5, 5), paired = FALSE, h = 0.2)
}
\arguments{
\item{p.cutoff}{default is 0.05.}

\item{group}{group set.}

\item{splot}{default is FALSE.}

\item{unitest}{t.test or wilcox.test.}

\item{pcorrect}{default is TRUE.}

\item{xlim}{xlim.}

\item{paired}{paired test in t.test and wilcox.test,default is FALSE.}

\item{h}{the height of group index,default is 0.2}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA,PLSDA,heatmap,
s plot of foldchange and volcano plot, also can calculate vip value.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
##create a folder for Shine demo
dir.create("Demo for Shine")
setwd("Demo for Shine")
##load the demo data
data(data, package = "Shine")
data(sample.info, package = "Shine")
##export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.info, "sample.info.csv", row.names = FALSE)
##run StaAnalysis
StaAnalysis(data = data,sample.info = sample.info,group = c("G","M"),
pcorrect = F)
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
