% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaAnalysis_SingleLine.R, R/StaAnalysis_SingleLine.R
\name{StaAnalysis_SingleLine}
\alias{StaAnalysis_SingleLine}
\title{StaAnalysis_SingleLine}
\usage{
StaAnalysis_SingleLine(data = NULL, sample.info = NULL, p.cutoff = 0.05,
  group = c("case", "control"), heatmap = FALSE, splot = FALSE,
  pcorrect = TRUE)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{p.cutoff}{default is 0.05.}

\item{group}{group set.}

\item{heatmap}{default is FALSE.}

\item{splot}{default is FALSE.}

\item{pcorrect}{default is TRUE.}

\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{group}{group set.}

\item{p.cutoff}{default is 0.05.}

\item{heatmap}{default is FALSE.}

\item{splot}{default is FALSE.}

\item{pcorrect}{default is TRUE.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA,PLSDA,heatmap,
s plot of foldchange and volcano plot, also can calculate vip value.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
data(data, package = "Shine")
data(sample.info, package = "Shine")
##create a folder for demo
dir.create("demo")
setwd("demo")
# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.info, "sample.info.csv", row.names = FALSE)
# Analysis process
StaAnalysis_SingleLine(data = "data.csv",sample.info = "sample.info.csv",group =c("S","P"),
p.cutoff = 0.05)
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
