% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{string.edges}
\alias{string.edges}
\title{Get a string-representation of igraph edges}
\usage{
string.edges(ig, rev = FALSE)
}
\arguments{
\item{ig}{An igraph object}

\item{rev}{Use true to return reversed edges}
}
\value{
A character vector of edges
}
\description{
Get a string-representation of igraph edges
}
\keyword{internal}
