% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy}
\alias{toy}
\title{Toy Data}
\format{
An eset with 100 variables and 30 samples:
\describe{
  \item{exprs(toy)}{A 100 x 30 matrix of multivariate Gaussian data}
  \item{pData(toy)}{A 30 x 1 matrix assigning subtypes of either A, B, or C}
  \item{fData(toy)}{A 30 x 1 matrix storing variable information}
}
}
\usage{
toy
}
\description{
An expression set object for data encoding a scale-free graph
}
\keyword{datasets}
