% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{TransformData}
\usage{
TransformData(
  data_name = "data.csv",
  name = "compound.name",
  boxplot_data = TRUE
)
}
\arguments{
\item{data_name}{data_position}

\item{name}{compound.name}

\item{boxplot_data}{default is TRUE}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to TransformData
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
