% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueCompound.R
\name{UniqueCompound}
\alias{UniqueCompound}
\title{unique compound}
\usage{
UniqueCompound(data)
}
\arguments{
\item{data}{a dataframe must include compound name and ID.}
}
\description{
unique compound
}
