% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.R
\name{Violinplot_Shine}
\alias{Violinplot_Shine}
\title{Violinplot_Shine}
\usage{
Violinplot_Shine(
  ppt_width = 9,
  font_size = 20,
  levels = c("Normal", "CHD", "MI")
)
}
\arguments{
\item{ppt_width}{default is 9}

\item{font_size}{font_size}

\item{levels}{order of group}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
