% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiClass.R
\name{BiClass}
\alias{BiClass}
\title{BiClass}
\usage{
BiClass()
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can do classify analysis using logistic,svm and randomforest
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
