% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{blanket.inform}
\alias{blanket.inform}
\title{Adds prior information where the blanket is lifted}
\usage{
blanket.inform(blanket, prior.edges, prior.blanket = NULL, val = 0.5)
}
\arguments{
\item{blanket}{A blanket matrix}

\item{prior.edges}{A matrix of prior edge probabilities}

\item{prior.blanket}{The blanket used in estimating prior edge probabilities}

\item{val}{Value for when a prior blanket is included}
}
\value{
A blanket matrix
}
\description{
Adds prior information where the blanket is lifted
}
