% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_Shine.R
\name{Boxplot_Shine}
\alias{Boxplot_Shine}
\title{Boxplot_Shine}
\usage{
Boxplot_Shine(
  font_size = 20,
  stat_type = "parametric",
  data_name = "data.csv",
  plot.type = "box",
  pairwise.display = "significant",
  p.adjust.method = "fdr",
  ylab = "log2 metabolites abundance",
  levels = c("M1", "M2", "M3")
)
}
\arguments{
\item{font_size}{font_size}

\item{stat_type}{parametric}

\item{data_name}{data.csv}

\item{plot.type}{box}

\item{pairwise.display}{significant}

\item{p.adjust.method}{fdr}

\item{ylab}{name}

\item{levels}{order of group}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
