% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_Shine.R
\name{Boxplot_Shine}
\alias{Boxplot_Shine}
\title{Boxplot_Shine}
\usage{
Boxplot_Shine(
  font_size = 20,
  levels = c("NU", "HU", "Gout"),
  height = 7,
  width = 9
)
}
\arguments{
\item{font_size}{font_size}

\item{levels}{order of group}

\item{height}{7}

\item{width}{9}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
