% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_Shine.R
\name{Boxplot_Shine}
\alias{Boxplot_Shine}
\title{Boxplot_Shine}
\usage{
Boxplot_Shine(
  ppt_width = 9,
  ngroup = 3,
  levels = c("Normal", "Hyperuricemia", "Gout"),
  palette = c("BottleRocket1", "BottleRocket2", "Rushmore1", "Royal1", "Royal2",
    "Zissou1", "Darjeeling1", "Darjeeling2", "Chevalier1", "FantasticFox1", "Moonrise1",
    "Moonrise2", "Moonrise3", "Cavalcanti1", "GrandBudapest1", "GrandBudapest2",
    "IsleofDogs1", "IsleofDogs2")
)
}
\arguments{
\item{ppt_width}{default is 9}

\item{ngroup}{number of grouo}

\item{levels}{order of group}

\item{palette}{colour selection}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
