% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow.R
\name{build.workflow}
\alias{build.workflow}
\title{Generate a nextflow workflow}
\usage{
build.workflow(
  hierarchy,
  condition,
  eset,
  blanket = NULL,
  outdir = ".",
  iter = 5000,
  cores = 1
)
}
\arguments{
\item{hierarchy}{A string describing the hierarchy in dot language notation}

\item{condition}{The column in pData where subtypes are described}

\item{eset}{An expression set object}

\item{blanket}{A gene x gene constraint matrix}

\item{outdir}{A directory where the workflow will be generated}

\item{iter}{The number of iterations for the sampling algorithm}

\item{cores}{The number of cores for each network to use for parallel execution}
}
\description{
Generate a nextflow workflow
}
