% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorAnalysis.R
\name{CorAnalysis}
\alias{CorAnalysis}
\title{CorAnalysis}
\usage{
CorAnalysis(number.cex = 0.6, tl.cex = 0.6, number.digits = 4,
  adjust = c("holm", "bonferroni", "BH", "fdr", "none"))
}
\arguments{
\item{number.cex}{The cex parameter to send to the call to text when
writing the correlation coefficients into the plot.}

\item{tl.cex}{font size of label}

\item{number.digits}{indicating the number of decimal digits to be added
into the plot. Non-negative integer or NULL, default 4.}

\item{adjust}{What adjustment for multiple tests should be used}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can do correlation analysis
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
