% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxAnalysis.R
\name{CoxAnalysis}
\alias{CoxAnalysis}
\title{CoxAnalysis}
\usage{
CoxAnalysis()
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to do cox Proportional-Hazards Model
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
