% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureAnalysis.R
\name{FeatureAnalysis}
\alias{FeatureAnalysis}
\title{FeatureAnalysis}
\usage{
FeatureAnalysis(data = NULL, sample.info = NULL, zero.filter = FALSE,
  RSD.filter = FALSE, zero.check = TRUE)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{zero.filter}{default is FALSE,if the zero value exsit in your data,
make FALSE to TRUE.}

\item{RSD.filter}{default is FALSE,if the percentage of qc rsd larger than 0.3
more than 0.7,almost after normalization,make FALSE to TRUE.}

\item{zero.check}{default is TRUE.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate mz vs rt plot and QC distribution plot,
also can filter isotope,rsd and zero value.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
##create a folder for Shine demo
dir.create("Demo for Shine")
setwd("Demo for Shine")
#run FeatureAnalysis
FeatureAnalysis(data = data,sample.info = sample.info,RSD.filter = T,
zero.check = F)
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
