% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestAnalysis.R
\name{ForestAnalysis}
\alias{ForestAnalysis}
\title{ForestAnalysis}
\usage{
ForestAnalysis(group = c("case", "control"), ncol = 19)
}
\arguments{
\item{group}{group set.}

\item{ncol}{the columns of data}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can do forest analysis
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
