% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{fuzzy.mods}
\alias{fuzzy.mods}
\title{Classify fuzzy modules with quadratic disciminant analysis}
\usage{
fuzzy.mods(dat, mods, fn = bicor)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{fn}{A correlation function}
}
\value{
A list of fuzzy modules
}
\description{
Classify fuzzy modules with quadratic disciminant analysis
}
