% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{fuzzy.mods}
\alias{fuzzy.mods}
\title{Classify fuzzy modules with quadratic disciminant analysis}
\usage{
fuzzy.mods(dat, mods, fn = bicor)
}
\description{
Classify fuzzy modules with quadratic disciminant analysis
}
