% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{fuzzy.plot}
\alias{fuzzy.plot}
\title{Visualize fuzzy module classification}
\usage{
fuzzy.plot(dat, mods, mod, size = 2.5, resolution = 100, fn = bicor)
}
\description{
Visualize fuzzy module classification
}
