% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{fuzzy.predict}
\alias{fuzzy.predict}
\title{Classify fuzzy membership for a module with quadratic discriminant analysis}
\usage{
fuzzy.predict(dat, mods, mod, p = 0.5, fn = bicor)
}
\arguments{
\item{dat}{Expression data}

\item{mods}{A list of modules}

\item{mod}{A module name}

\item{p}{Posterior probability threshold}

\item{fn}{A correlation function}
}
\value{
The model and a dataframe of classifications
}
\description{
Classify fuzzy membership for a module with quadratic discriminant analysis
}
