% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Shine.R
\name{Heatmap_Shine}
\alias{Heatmap_Shine}
\title{Heatmap_Shine}
\usage{
Heatmap_Shine(colour = c("green", "black", "red"), a = -2, b = 2,
  d = 100, scale_row = TRUE, size_row = 10, size_col = 8,
  fontsize = 10, border = NA, cluster_cols = FALSE,
  group = c("case", "control"))
}
\arguments{
\item{colour}{colour selection.}

\item{a}{lowerlimt of colour.}

\item{b}{toplimit of colour}

\item{d}{default 100}

\item{scale_row}{scale or not}

\item{size_row}{the font size of row}

\item{size_col}{the font size of col}

\item{fontsize}{the font size}

\item{border}{show boder or not}

\item{cluster_cols}{default is FALSE}

\item{group}{group info.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate heatmap
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
