% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.mer}
\alias{model.mer}
\title{Simulate network via modular erdos-renyi model}
\usage{
model.mer(p = 300, m = 6, m.prob = 0.07, n.inter = 1, ...)
}
\description{
Simulate network via modular erdos-renyi model
}
