% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.mpa}
\alias{model.mpa}
\title{Simulate network via modular preferential attachment model}
\usage{
model.mpa(
  p = 300,
  m = 6,
  q.hub = 0.95,
  m.links = 6,
  power = 1.7,
  z.appeal = 1,
  ...
)
}
\arguments{
\item{p}{The number of vertices in the graph}

\item{m}{The number of modules}

\item{q.hub}{The quantile for degree cutoff for defining modular hubs}

\item{m.links}{The number of links between modules}

\item{power}{The power of the preferential attachment}

\item{z.appeal}{The attractiveness of the vertices with no adjacent edges}

\item{...}{Additional arguments passed to /code{igraph::sample_pa}}
}
\value{
An igraph object
}
\description{
Simulate network via modular preferential attachment model
}
