% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model.plot}
\alias{model.plot}
\title{Plot model}
\usage{
model.plot(ig, seed = 1, colors = FALSE)
}
\arguments{
\item{ig}{An igraph object}

\item{seed}{A number to seed random layouts}
}
\value{
A graph visualization
}
\description{
Plot model
}
