% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayBarplot.R
\name{PathwayBarplot}
\alias{PathwayBarplot}
\title{pathway enrichment barplot}
\usage{
PathwayBarplot(row = 30)
}
\arguments{
\item{row}{the rows be selected to plot,default is 30.}
}
\description{
pathway enrichment barplot
}
