% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayBarplot.R
\name{PathwayBarplot}
\alias{PathwayBarplot}
\title{PathwayBarplot}
\usage{
PathwayBarplot(font_size = 26, color = c("#3399FF", "#FF3333"))
}
\arguments{
\item{font_size}{font_size}

\item{color}{color}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to draw BoxPlot
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
