% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayBarplot.R
\name{PathwayBarplot}
\alias{PathwayBarplot}
\title{pathway enrichment barplot}
\usage{
PathwayBarplot(data)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}
}
\description{
pathway enrichment barplot
}
