% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayEnrich.R
\name{PathwayEnrich}
\alias{PathwayEnrich}
\title{PathwayEnrich}
\usage{
PathwayEnrich(
  specias_pathway_database = hsa.kegg.pathway,
  specias_compound_database = hsa_compound_ID,
  font_size = 20,
  data_name = "data_pathway.csv",
  color = c("#3399FF", "#FF3333"),
  All_ID = FALSE,
  seed = TRUE,
  grade1_2 = FALSE
)
}
\arguments{
\item{specias_pathway_database}{the specia of sample}

\item{specias_compound_database}{the specia of sample}

\item{font_size}{20}

\item{data_name}{the data_name}

\item{color}{color}

\item{All_ID}{default is FALSE}

\item{seed}{default is TRUE}

\item{grade1_2}{default is FALSE}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to do PathwayAnalysis
}
\author{
Xia Shen
\email{qq951633542@163.com}
}
