% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayEnrich.R
\name{PathwayEnrich}
\alias{PathwayEnrich}
\title{PathwayEnrich}
\usage{
PathwayEnrich(specias_pathway_database = c(hsa.kegg.pathway,
  mmu.kegg.pathway), specias_compound_database = c(hsa_compound_ID,
  mmu_compound_ID), keggmap = TRUE)
}
\arguments{
\item{specias_pathway_database}{the specia of sample}

\item{specias_compound_database}{the specia of sample}

\item{keggmap}{generate file for keggmap}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can match differetiate metabolites in differentiate pathway.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
