% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMatch.R
\name{PathwayMatch}
\alias{PathwayMatch}
\title{PathwayMatch}
\usage{
PathwayMatch(data = NULL, sample.info = NULL, group = c("case",
  "control"), pcorrect = TRUE, pathway.name = pathway)
}
\arguments{
\item{data}{a dataframe from MetDNA named Quantitative.pathway.metabolite.result.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{group}{group set.}

\item{pcorrect}{default is TRUE.}

\item{pathway.name}{it is the pathway you want to match.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can match differetiate metabolites in differentiate pathway.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
library(Shine)
data("hsa.kegg.pathway",package="Shine")
data("mmu.kegg.pathway",package="Shine")
load(h)
load(m)
pathway<-hsa.kegg.pathway$'Cysteine and methionine metabolism'
PathwayMatch(data = data,sample.info = sample.info,group = c("S","P"),
pcorrect = F,pathway.name = "Cysteine and methionine metabolism")
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
