% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMatch.R
\name{PathwayMatch}
\alias{PathwayMatch}
\title{PathwayMatch}
\usage{
PathwayMatch(group = c("case", "control"))
}
\arguments{
\item{group}{group set.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can match differetiate metabolites in differentiate pathway.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
library(Shine)
PathwayMatch(group = c("S","P"))
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
