% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMatch.R
\name{PathwayMatch}
\alias{PathwayMatch}
\title{PathwayMatch}
\usage{
PathwayMatch(data = NULL, sample.info = NULL, group = c("case",
  "control"), pcorrect = TRUE, pathway.name = pathway)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{group}{group set.}

\item{pcorrect}{default is TRUE.}

\item{pathway.name}{it is the pathway you want to match.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can match differetiate metabolites in differentiate pathway.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
