% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Shine.R
\name{PCA_Shine}
\alias{PCA_Shine}
\title{PCA_Shine}
\usage{
PCA_Shine(ind = FALSE, ellipse = FALSE, both = FALSE, neither = TRUE)
}
\arguments{
\item{ind}{default is FALSE.}

\item{ellipse}{default is FALSE.}

\item{both}{default is FALSE.}

\item{neither}{default is TRUE.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
