% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_Shine.R
\name{PLSDA}
\alias{PLSDA}
\title{PLSDA_Shine}
\usage{
PLSDA(
  ind = FALSE,
  ellipse = FALSE,
  both = FALSE,
  neither = TRUE,
  group = c("case", "control"),
  multiclass = FALSE
)
}
\arguments{
\item{ind}{default is FALSE.}

\item{ellipse}{default is FALSE.}

\item{both}{default is FALSE.}

\item{neither}{default is TRUE.}

\item{group}{group info.}

\item{multiclass}{multiclass plsda}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA,PLSDA,heatmap,
s plot of foldchange and volcano plot, also can calculate vip value.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
 #the data must be named as 'data.csv' and 'sample.info.csv'
PLSDA()
}
}
\author{
Xia Shen
\email{qq951633542@163.com}
}
