% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_s3.R
\name{read_s3}
\alias{read_s3}
\title{Read Data from S3 Storage}
\usage{
read_s3(path, bucket, format = c("auto", "qs", "rds", "csv", "parquet"))
}
\arguments{
\item{path}{Character string specifying the path to the file in S3}

\item{bucket}{Character string specifying the S3 bucket name}

\item{format}{Character string specifying the file format. One of "auto", "qs", "rds", "csv", or "parquet"}
}
\value{
The data read from the S3 file in the specified format
}
\description{
Read Data from S3 Storage
}
