% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaAnalysis.R
\name{StaAnalysis}
\alias{StaAnalysis}
\title{StaAnalysis}
\usage{
StaAnalysis(data = NULL, sample.info = NULL, p.cutoff = 0.05,
  group = c("case", "control"), heatmap = FALSE, splot = FALSE,
  unitest = c("t.test", "wilcox.test"), pcorrect = TRUE)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{p.cutoff}{default is 0.05.}

\item{group}{group set.}

\item{heatmap}{default is FALSE.}

\item{splot}{default is FALSE.}

\item{unitest}{t.test or wilcox.test.}

\item{pcorrect}{default is TRUE.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA,PLSDA,heatmap,
s plot of foldchange and volcano plot, also can calculate vip value.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
##create a folder for Shine demo
dir.create("Demo for Shine")
setwd("Demo for Shine")
##load the demo data
data(data, package = "Shine")
data(sample.info, package = "Shine")
##export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.info, "sample.info.csv", row.names = FALSE)
##run StaAnalysis
StaAnalysis(data = data,sample.info = sample.info,group = c("G","M"),
pcorrect = F)
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
