% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaAnalysis.R
\name{StaAnalysis}
\alias{StaAnalysis}
\title{StaAnalysis}
\usage{
StaAnalysis(p.cutoff = 0, group = c("case", "control"),
  splot = FALSE, unitest = c("t.test", "wilcox.test"),
  specias_pathway_database = c(hsa.kegg.pathway, mmu.kegg.pathway),
  specias_compound_database = c(hsa_compound_ID, mmu_compound_ID),
  pcorrect = TRUE, xlim = c(-3, 3), paired = FALSE, h = 0.2,
  PCA = FALSE)
}
\arguments{
\item{p.cutoff}{default is 0.05.}

\item{group}{group set.}

\item{splot}{default is FALSE.}

\item{unitest}{t.test or wilcox.test.}

\item{specias_pathway_database}{the specia of sample}

\item{specias_compound_database}{the specia of sample}

\item{pcorrect}{default is TRUE.}

\item{xlim}{xlim.}

\item{paired}{paired test in t.test and wilcox.test,default is FALSE.}

\item{h}{the height of group index,default is 0.2}

\item{PCA}{draw PCA plot or not}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA,PLSDA,heatmap,
s plot of foldchange and volcano plot, also can calculate vip value.
}
\examples{
\donttest{
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
