% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaAnalysis.R
\name{StaAnalysis}
\alias{StaAnalysis}
\title{StaAnalysis}
\usage{
StaAnalysis(
  p.cutoff = 0,
  group = c("case", "control"),
  splot = FALSE,
  colv = c("SpringGreen3", "grey", "Firebrick1"),
  unitest = c("t.test", "wilcox.test"),
  pcorrect = TRUE,
  paired = FALSE,
  h = 0.2,
  PCA = FALSE,
  heatmap = FALSE
)
}
\arguments{
\item{p.cutoff}{default is 0.05.}

\item{group}{group set.}

\item{splot}{default is FALSE.}

\item{colv}{the colour of volcano}

\item{unitest}{t.test or wilcox.test.}

\item{pcorrect}{default is TRUE.}

\item{paired}{paired test in t.test and wilcox.test,default is FALSE.}

\item{h}{the height of group index,default is 0.2}

\item{PCA}{draw PCA plot or not}

\item{heatmap}{heatmap}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA,PLSDA,heatmap,
s plot of foldchange and volcano plot, also can calculate vip value.
}
\examples{
\donttest{
StaAnalysis()
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
