% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaAnalysis.R
\name{StaAnalysis}
\alias{StaAnalysis}
\title{StaAnalysis}
\usage{
StaAnalysis(data = NULL, sample.info = NULL, group = c("case", "control"))
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{group}{group set}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate PCA,PLSDA,heatmap,
s plot of foldchange and volcano plot, also can calculate vip value.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
data(data, package = "Shine")
data(sample.info, package = "Shine")
##create a folder for demo
dir.create("demo")
setwd("demo")
# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.info, "sample_info.csv", row.names = FALSE)
# Analysis process
StaAnalysis(data = "data.csv",sample.info = "sample.info.csv",group =c("S","P"))
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
