% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{TransformData}
\usage{
TransformData(
  group = "Normal",
  data_name = "lasso_variable.csv",
  name = "compound.name",
  boxplot_data = TRUE,
  roc = FALSE
)
}
\arguments{
\item{group}{the group you want to remove}

\item{data_name}{data_position}

\item{name}{compound.name}

\item{boxplot_data}{default is TRUE}

\item{roc}{default is FALSE}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to TransformData
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
