% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{TransformData}
\usage{
TransformData(group = "N", data_position = 1, info_position = 2,
  name = "compound.name", boxplot_data = FALSE)
}
\arguments{
\item{group}{the group you want to remove}

\item{data_position}{data_position}

\item{info_position}{info_position}

\item{name}{compound.name}

\item{boxplot_data}{default is FALSE}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function to TransformData
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
