% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueCompound.R
\name{UniqueCompound}
\alias{UniqueCompound}
\title{unique compound}
\usage{
UniqueCompound(data)
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}
}
\description{
unique compound
}
