% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{volcano}
\alias{volcano}
\title{volcano}
\usage{
volcano(p.cutoff = 0, group = c("case", "control"), pcorrect = TRUE,
  singleline = TRUE, xlim = c(-3, 3), fc.cutoff = 0.41,
  doubleline = FALSE, unitest = c("t.test", "wilcox.test"),
  paired = FALSE, h = 0.2)
}
\arguments{
\item{p.cutoff}{default is 0.05.}

\item{group}{group set.}

\item{pcorrect}{default is TRUE.}

\item{singleline}{default is FALSE.}

\item{xlim}{xlim.}

\item{fc.cutoff}{default is 1}

\item{doubleline}{default is TRUE.}

\item{unitest}{t.test or wilcox.test.}

\item{paired}{paired test in t.test and wilcox.test,default is FALSE.}

\item{h}{the height of group index,default is 0.2}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate volcano plot.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
