% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{volcano}
\alias{volcano}
\title{volcano}
\usage{
volcano(data = NULL, sample.info = NULL, p.cutoff = 0.05,
  group = c("case", "control"), pcorrect = TRUE, singleline = TRUE,
  xlim = c(-5, 5), doubleline = FALSE, unitest = c("t.test",
  "wilcox.test"))
}
\arguments{
\item{data}{a dataframe include name,mz,rt and isotope columns,
the rest of all are sample and QC columns.}

\item{sample.info}{a dataframe include sample.name,injection.order,
class,batch and group columns.}

\item{p.cutoff}{default is 0.05.}

\item{group}{group set.}

\item{pcorrect}{default is TRUE.}

\item{singleline}{default is FALSE.}

\item{doubleline}{default is TRUE.}

\item{unitest}{t.test or wilcox.test.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
a function can generate volcano plot.
}
\examples{
\donttest{
##---- Be sure the format of data and sample.info is correct!! ----
}
}
\author{
Shine Shen
\email{qq951633542@163.com}
}
