% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_s3.R
\name{write_s3}
\alias{write_s3}
\title{Write Data to S3 Storage}
\usage{
write_s3(data, bucket, format = c("qs", "rds", "csv", "parquet"), name = NULL)
}
\arguments{
\item{data}{Data to be written to S3}

\item{bucket}{Character string specifying the S3 bucket name}

\item{format}{Character string specifying the file format. One of "qs", "rds", "csv", or "parquet"}

\item{name}{Optional character string specifying the output file name}
}
\value{
Nothing. Writes data to S3 as a side effect.
}
\description{
Write Data to S3 Storage
}
